First cut at creating the ant task to execute an xslt stylesheet to convert a RDF/RDFS file to user readable HTML. To install this, please do the following:
1. cd ant-task-plugins
2. mvn clean install

The ant task currently takes three parameters:

"input" - The name of the input file to convert
"output" - The name of the output file
"sheet" - Name of the style sheet to use for the conversion

To make it easier to use, you might want to create a settings.xml file in ~/.m2
with the following entry:

<settings>
  <pluginGroups>
      <pluginGroup>info.anumana.plugins</pluginGroup>
  </pluginGroups>
</settings>

If you do not do this above step, you will need to type in info.anumana.plugins
before every access to the ant-tasks list.

To convert the file foaf.rdf to foaf.html with the stylesheet rdfs2html.xsl,
you can use the following command:

mvn -Dinput=foaf.rdf -Dsheet=rdfs2html.xsl -Doutput=foaf.html ant-tasks:rdfs2html

Note: The ant xslt task is extremely capable of processing directories
containing input files, and we will probably move towards that as the structure
and content of where the RDF/RDFS files are stored becomes clearer.



git-svn-id: http://svn.topazproject.org/svn/head@18 48088309-b40f-0410-b149-b14953253675