Added the following tasks:   fedora-install : to unpack and install topaz/anumana fedora version   fedora-start   : starts mckoi and fedora.   fedora-stop    : stops fedora and mckoi
You can run these tasks from anywhere like this:
  mvn ant-tasks:fedora-install
  mvn ant-tasks:fedora-start
  mvn ant-tasks:fedora-stop

  ... provided you have configured this shortcut in your settings.xml file:

  <pluginGroups>
      <pluginGroup>info.anumana.plugins</pluginGroup>
  </pluginGroups>

Default for the install directory is ~/fedora-install. It can be changed by giving a -DFEDORA_INSTALL_DIR=<a dir> option. You also need to pass -DFEDORA_HOME=<a dir>/fedora-2.1.1 to the fedora-start and fedora-stop commands if you change the defaults. 

Nuisance warning:
   ... when doing fedora-start command, once fedora has started up, you have to do a Ctl-C to kill the mvn command. The fedora service will not be killed. The shell that runs this command for some reason does not return.

NOT TESTED ON WINDOWS.



git-svn-id: http://svn.topazproject.org/svn/head@33 48088309-b40f-0410-b149-b14953253675