Replace gandalf with gandalf.topazproject.org in maven poms.
Ronald indicated that it took a while to get the DNS entry for
gandalf.topazproject.org and that current developers all have
"gandalf" in their /etc/hosts. But this should make it easier
for new developers. 

I wasn't sure exactly how to test the <scm> and 
<distributionManagement> sections, but I don't see how this
change would cause problems for those sections either.


git-svn-id: http://svn.topazproject.org/svn/head@141 48088309-b40f-0410-b149-b14953253675