Added a simple archetype for creating topaz webservices. Usage:   mvn archetype:create -DarchetypeGroupId=org.topazproject.archetypes -DarchetypeArtifactId=webservice -DarchetypeVersion=1.0-SNAPSHOT -DgroupId=<svc-group-id> -DartifactId=<svc-name> -DpackageName=<java-package>
where <svc-group-id> is typically ''org.topazproject.ws'', <svc-name> is the name of the service,
and <java-package> is the java package the code will be put in (typically
''org.topazproject.ws.<svc-name>''). The <svc-name> is used in some comments (e.g. "this class
defines the <svc-name> service"); the name with the first letter capitalized is used to create
the class names and for comments in the pom.xml's.

Example:
  mvn archetype:create -DarchetypeGroupId=org.topazproject.archetypes -DarchetypeArtifactId=webservice -DarchetypeVersion=1.0-SNAPSHOT -DgroupId=org.topazproject.ws -DartifactId=messages -DpackageName=org.topazproject.ws.messages

There are some limitations in the current maven archetype stuff, which lead to the following:
 * After running ''mvn archetype:create ...'', find all XYZ*.java files and rename to
   <SvcName>*.java (where <SvcName> is the capitalized <svc-name> from the command, e.g.
   ''Messages'')
 * Rename ''XYZ.xml'', and replace the ''-some-method-'' marker appropriately; adjust the
   ''PDPConfig.xml'' to match the filename.
 * Because you can't pass arbitrary parameters to the archectype stuff, in the command above, the
   -DartifactId has been slight misused to carry the service-name instead; so you might want to
   edit the top-level pom.xml to change the artifactId.

Now you can fill in the service definition in the <SvcName>.java and the implementation thereof in
the <SvcName>Impl.java class; you'll also need to update ''PDPConfig.xml'' with the list of policy
files to use and update those policy files with the proper xacml definitions.



git-svn-id: http://svn.topazproject.org/svn/head@153 48088309-b40f-0410-b149-b14953253675