Use a fedora-server.jar instead of untar and build against the fedora distribution tar file.
Using javac from within the ant-run requires declaring sun-jdk as a dependency and also making an assumption about the location of tools.jar. The java.home in maven is the jre home and not the jdk home. So the relative locations are different on a mac vs linux or windows. This means there isn't a clean way to access javac from ant-run.

So the workaround is to create a jar file from fedora/WEB-INF/classes directory and deploy that on our maven repository on gandalf under the groupId info.fedora.

The resource-indexer pom declares fedora-server jar as a dependency and has become quite simple as a result. 



git-svn-id: http://svn.topazproject.org/svn/head@154 48088309-b40f-0410-b149-b14953253675