Re-ordered the repository list (ibilio followed by codehaus followed by gandalf).
So after you update to this new pom, suggest that you delete your 
~/.m2/repository (or definitely the ~/.m2/repository/plexus) and 
let maven re-download the required artifacts.

Gory details on why this re-ordering was done:

I couldn't build the archetype/webservices after the change to 1.0-alpah-3 
for archetype-plugin in r187. The groupId 'plexus' is present in both codehaus 
and at ibiblio. The ones at codehaus are older(obsolete?) versions but with 
the same artifactId and version number. The culprit pom that broke my build was: 

http://dist.codehaus.org/plexus/poms/plexus-components-1.0.pom

The correct one should be:

http://www.ibiblio.org/maven2/plexus/plexus-components/1.0/plexus-components-1.0.pom



git-svn-id: http://svn.topazproject.org/svn/head@190 48088309-b40f-0410-b149-b14953253675