Added ibiblio and codehaus to pluginRepositories list.
mvn automatically tries ibiblio for plugins. However it does so
after it tries the ones listed in the pom. Since the repository
on gandalf only contains the ones that are not present in ibilio,
reversing the search order by explicitly specifying ibiblio.

codehaus is added since plugins like xdoclet (used by plosone) is 
not present at ibiblio.



git-svn-id: http://svn.topazproject.org/svn/head@191 48088309-b40f-0410-b149-b14953253675