Prevent the bogus redirect to cas login when a proxy ticket validation fails. Instead return an un-authorized(401) error with a BASIC authentication challenge.
Proxies have no way to fill out a login form. So the redirect is meaningless.
Returning a 401 on the otherhand can be used by proxies to may be re-get a
CAS ticket and retry the request.

May be we should consider sending a different challenge back to the proxies instead of
WWW-Authenticate: Basic realm="a-realm".

All these are only applicable to proxies. For direct user-agent requests, the 
redirect to CAS login will proceed as before. (eg. for AJAX/browser requests
for fedora URLs). 



git-svn-id: http://svn.topazproject.org/svn/head@222 48088309-b40f-0410-b149-b14953253675