Fixed issue with non-guaranteed execution order of plugins in same phase
The java2wsdl and antrun task to copy the output of java2wsdl were both
running in the process-classes phase. The copied output was used by the
maven-jar-plugin in the packaging phase to put the generated .wsdl in
the root of the jar file. And, unfortunately, there were no reasonable
phases between process-classes and packaging to shove the ant-task and
the maven-jar-plugin is not very configurable. The best solution was to
have java2wsdl put the .wsdl into target/classes directly instead of 
into the default target/generated-sources/axistools/java2wsdl directory.

(I hope I'm not missing some reason the .wsdl wants to be in the
generated-sources directory too.)

The problem I was hitting was this pom.xml was inheriting new build 
information from a new profile I have yet to check-in. And when it did
so maven reordered the plugin execution order so the ant-task tried to
run before the java2wsdl -- something that obviously errored out. It is
clear that maven does not honor the order plugins are placed in a
pom.xml when determining what order to run things within a given phase.

re #63 makes inherited profiles possible


git-svn-id: http://svn.topazproject.org/svn/head@241 48088309-b40f-0410-b149-b14953253675