re #63 Re-enable integration tests through the it-helper profile
The it-helper profile is defined in head/pom.xml and topazproject/webservices/
*/ws-client-impl/pom.xml. If "mvn -Pit-helper integration-tests" is run in
any of the ws-client-impl modules, integration tests will be run. However, if
the profile is off (default), integration tests are not run and everything
builds normally.

Unfortunately, running the it-helper profile outside of the specific sub-
modules will fail due to the annoying way maven decises when and how to 
compute the pom to run. So don't try it! This is phase 1. Phase 2 of the
integration test stuff will address that.

Integration test code does assume that fedora and CAS are running.

Integration test code has been moved from src/test to src/it. This is a
pattern, not necessarily the best pattern given difficulties getting maven
to use this directory as intended. Time will tell.

The it-helper profile defines the following:
 * testResources = src/it/resources AND src/test/resources
   (both needed or unit-tests may break - it is a compromise)
 * src/it/java added to test sources via build-helper-maven-plugin
 * surefire unit-tests turned off
 * surefire integration-tests turned on
 * A default configuration for cargo/jetty is defined such that all
   a sub-module needs to do is define that it wants to use the plugin
   and all defaults will be pulled in. (This is so that running with
   this profile in non ws-client-impl modules may still work somewhat.)

=== How to test ===
for dir in topazproject/webservices/*/ws-client-impl; \
  do (cd $dir; mvn -Pit-helper integration-test); done


git-svn-id: http://svn.topazproject.org/svn/head@261 48088309-b40f-0410-b149-b14953253675