fixes #63 Integration framework for topaz
Please see IntegrationTesting wiki page.

Integration tests are those that assume a container is present. The tests
should go into one of two places:
 * <your-module>/src/it
 * <integrationtests-module>/src/it

Tests are run by:
{{{
  cd <integrationtests-module>
  mvn -Pit-startenv clean integration-test
}}}

Topaz and plosone will each have their own <integrationtests-module>. 
Both will be fairly similar, but it is likely different tools may be used.
(i.e. plosone might choose to use cactus). That said, head/topazproject/
integrationtests is a pattern that should easily be replicated for plosone.

Maven2 is a bit immature and multiple sources indicated that the best way
to implement integration testing at this time was to create a separate
module as I've done. However, there is quite a bit of maintenance over-
head required to ensure this module keeps working -- primarily, it must
have dependencies on all modules and modules those modules depend on to
test in its dependency list. This is duplicated effort and there is 
currently no way around it.

The integrationtests module looks for integration tests in **/src/it in
its peer directories and all their subdirectories.

To run tests one-module at a time, please see [261].

NB: If the integration-test stage fails, post-integration-test does not
get run. Vincent sent me a possible solution and Ronald and I discussed
writing a shell script to integrate into continuum.

=== How to test ===
{{{
  cd head
  mvn clean install
  cd topazproject/integrationtests
  mvn -Pitstartenv clean integration-test
}}}
If any of the tests fail, you may need to shutdown manually:
{{{
  mvn ant-tasks:fedora-stop
  mvn ant-tasks:ecqs-stop
}}}



git-svn-id: http://svn.topazproject.org/svn/head@298 48088309-b40f-0410-b149-b14953253675