Accept dynamic-typing for user-id in policies. (User-id is represented in xacml policies by urn:oasis:names:tc:xacml:1.0:subject:subject-id.) This modification allows PAPs to author policies that refernce the user-id by any type (eg. xsd:string or xsd:AnyURI).
An error in creating the value (eg. URISyntaxException) will result in a 
ParsingError exception and this in-turn will cause the PDP to return an 
'indeterminate' evaluation status and this in-turn will cause a 
Deny-Biased PEP to deny access. Therefore the policy author must choose
the type carefully to avoid any un-intentional access-denials.. 



git-svn-id: http://svn.topazproject.org/svn/head@307 48088309-b40f-0410-b149-b14953253675