Fix (sorta) problem with mvn -Pit-helper install; mvn install
Added a post-integration-test that removes target/test-classes
so that future runs of ''mvn install'' without the ''it-helper''
profile will not error out.

NB: If you run ''mvn -Pit-helper integration-test'' or you run
''mvn -Pit-helper install'' and your integration tests fail,
the post-integration-test phase will not get run the
target/test-classes directory will still be around to
potentially screw up runs that are not run with the it-helper
profile.

Spent at least an hour trying to get stuff built into it-classes
(or using antrun to move to it-classes), but in the end, was
unable to get maven-surefire-plugin to use the stuff from
it-classes anyway. Futile effort.

=== How to test ===
Go to a ws-client-impl directory of a webservices. Run
{{{mvn -Pit-helper clean install; mvn install}}}.

If you get no errors, you're fine.


git-svn-id: http://svn.topazproject.org/svn/head@327 48088309-b40f-0410-b149-b14953253675