Fix breakage introduced in rev [328] related to trippi.jar.
The trippi stuff is a mess. On sourceforge there are two versions, 0.9.9
for kowari 1.0.5, and 1.0.0 for kowari 1.1.0-pre2. However, what's shipped
with the binary versions of fedora is some other version trippi, ostensibly
1.0.2 according to the Trippi.properties file in the jar. That version is
for kowari 1.0.5, and hence ought to be the same as 0.9.9, but it isn't.
Furthermore, it's not clear where the source for this 1.0.2 version is -
it's certainly not part of the fedora src package.

So I've proceeded to patch the trippi supplied with fedora as follows. The
diffs between 0.9.9 and 1.0.0 show that only a few classes changed. Of
those that changed, only 3 are in 1.0.2. And futhermore, luckily, those
three are identical in the 0.9.9 and 1.0.2 versions, meaning it should be
safe to replace them with the 1.0.0 versions. So this is what the
re-packaging now does.


git-svn-id: http://svn.topazproject.org/svn/head@334 48088309-b40f-0410-b149-b14953253675