Added the ability to use ItqlQueryFunction as an abstract Function (along with its current usage as a function returning a bag of xsd:strings.)
When used as an abstract-function it looks for the attribute 'ReturnType' in
<Apply> node to figure out the return type.

Abstract function usage can be configured in the PDP config file
with a function identifier to use in policies. eg:

{{{

PDP Config file:

  <functionFactory name="func" useStandardFunctions="true">
    <condition>
      <abstractFunction class="org.topazproject.xacml.cond.ItqlQueryFunctionProxy"
        identifier="topaz:itql-abstract-function"/>
    </condition>
  </functionFactory>

Policy Usage:

   <Apply FunctionId ="topaz:itql-abstract-function"
          ReturnType="http://www.w3.org/2001/XMLSchema#anyURI">
      ...
      ...
   </Apply>

}}}

In addition to the above change, there are numerous other small fixes
as a result of finally doing some testing on it.

NOTE: If the xacml policies only need the xsd:string return type, then
      the regular function usage is recommended.



git-svn-id: http://svn.topazproject.org/svn/head@348 48088309-b40f-0410-b149-b14953253675