Generates topaz javadocs
This is not the final solution, but I thought the docs might be useful to
others and this works for now. It is a copy of all the stuff I did in
head/docs, but it requires the same set of <dependency> directives and it
is painful enough maintaining those in two places as it is - to add a third
place is unreasonable. (Sorry, it does not generate plosone java docs.)

To run, do:
{{{
cd head/topazproject; mvn generate-sources
cd head/topazproject/integrationtests; mvn site-deploy
}}}

Docs will end up in /tmp/site. This can be changed by setting the
<topazproject.site.url> in your settings.xml. That's the way it will be
done in continuum.

Running the generate-sources target will allow javadoc generation for
client services (which can be very useful).

re #91 and #93


git-svn-id: http://svn.topazproject.org/svn/head@362 48088309-b40f-0410-b149-b14953253675