Recognize the role of a mediator for creating annotations.
Mediator is an opaque attribute as far as the annotations service is 
concerned. Cooperating applications may choose to allocate unique
identifiers for partitioning the set of annotations stored by the service.
This may be an application-identifier or other schemes that an application
may choose to use.

listAnnotations now allows filtering by annotations created via a mediator.

Also fixed a bug in listing all annotations in any admin state.



git-svn-id: http://svn.topazproject.org/svn/head@446 48088309-b40f-0410-b149-b14953253675