fixes #117 Add email generation to alerts service
To use, set two user preferences ("alertsCategories" and "alertsEmailAddress")
and call alerts.startUser(). To stop sending alerts, remove the preferences
and call alerts.clearUser(). (startUser/clearUser ''recommendations'' may be
removed in the future.)

This is mostly code complete. There are a few more tasks (for which tickets
will be opened):
 * Run the service on a configurable schedule (#118)
 * More integration tests (See comment at top of AlertsServiceTest.java)
 * Improve email.xsl to generate better looking emails
 * Make relay-server/port configurable
 * Make startUser() call optional
 


git-svn-id: http://svn.topazproject.org/svn/head@447 48088309-b40f-0410-b149-b14953253675