Addresses #72: insert all meta-data directly into the triple-store.
However, due to the incompleteness of the fix for #10 in rev [469] we
also store that data into the DC and RELS-EXT streams of Fedora, i.e. we
explicitly store the meta-data in two places. Once #10 is fixed properly
we can (and should) stop populating the DC and RELS-EXT streams, at which
point #72 can be closed.


git-svn-id: http://svn.topazproject.org/svn/head@470 48088309-b40f-0410-b149-b14953253675