Fix #103. For services with renewable credentials, added a mechanism to re-authenticate and retry failed operation that failed due to expired credentials. The changes are:   * Added a ReAuthenticating stub factory   * Added an auth fault detection mechanism   * Added a renew() to ProtectedService   * Created concrete implementations of ProtectedService
All client stubs and stub factories (itql, fedora etc.) are
also updated to make use of the re-authenticating. 

fedora-client tests are moved to src/it and contains tests for re-auth.

Note that the fedora-uploader re-auth does not make use of the StubFactory
since uploader is not implemented as a client stub. So the retry is inlined in it.



git-svn-id: http://svn.topazproject.org/svn/head@482 48088309-b40f-0410-b149-b14953253675