Closes #88: removed the readers stuff from the profile, and instead rely on the permissions service and standard XACML policies for access-control management. A set of permissions, one for each field in the profile, is provided for reading the profile; a single permission still governs the setting of the profile.
The tests have been adjusted, and some basic setups for the access-control
tests done, but the actual tests can't be performed until a framework for
logging in and switching between users in tests is set up.

On a performance note: a profile retrieval results in as many check-access
calls (and hence XACML policy evaluations, and hence iTQL queries) as there
are fields in the profile (currently 10). We probably need to introduce
some caching somewhere at some point.


git-svn-id: http://svn.topazproject.org/svn/head@527 48088309-b40f-0410-b149-b14953253675