re #122 Test environment for log4j - patch a release, not SNAPSHOT
This eliminates dependency on codehaus snapshot repository and means
that we are not subject to changes or build problems introduced to the
current snapshot. We are fixed at maven2-cargo-plugin 0.2 / cargo 0.8.

I built our 0.2 patch and deployed it to gandalf. Will checkin necessary
stuff for that later. Basically I just grabbed tags/0.2-maven2, applied
the same patch as for 0.3-SNAPSHOT, but deleted the unit tests since
they will fail due to subtle API changes in the SNAPSHOT.


git-svn-id: http://svn.topazproject.org/svn/head@541 48088309-b40f-0410-b149-b14953253675