fixes #118 send alerts on a schedule
Mostly this just removes web.xml in favor of base-ws-webapp's web.xml per [427].
That was the only thing missing to making this work.

=== How to test ===
Not sure how to write an integration test for this. So just modify
alerts/ws-webapp/src/main/webapp/WEB-INF/topaz-config.xml to send alerts
every minute. (Set cron to "0 * * * * ?".) Run the integration tests and
notice that every minute on the minute, an attempt is made to send out all
alerts. Logging level of INFO is required. (It is possible depending on
timing that doing this will break the integration test as the job will
send out an alert that an integration test was expecting to send out.
It'd be fairly rare - like amost never, however, for this to happen.)


git-svn-id: http://svn.topazproject.org/svn/head@542 48088309-b40f-0410-b149-b14953253675