Closes #154: created a generic exception handler for the soap-binding-impl's which flattens exception chains so that the exception message returned to the client contains the messages from all the chained exceptions.
The handler also handles logging of the exceptions.

Additionally, all NoSuchXYZIdExceptions and DuplicateXYZIdExceptions have
been given a constructor which takes a message, and the id-only constructor
will generate a default message containing the id (so that toString() and
printStackTrace() show the id).

There is a caveat introduced by this change regarding the exceptions which
is documented in the ExceptionUtils.flattenException method.


git-svn-id: http://svn.topazproject.org/svn/head@646 48088309-b40f-0410-b149-b14953253675