Temporary hack to fix javadocs/maven report broken by [687]
Java javadoc message is very obtuse and doesn't make much sense to me.
Even running maven -e javadoc:javadoc chops of the exception trace
generated by javadoc -- and it makes it seem like the subsequent line
is the source of the error when I don't think it is. It may even be
more of a javadoc/maven issue -- not sure.

I don't understand why [687] breaks the javadocs, but I don't think
it is worth spending time on at the moment. Needs to be fixed
properly later. re #178.


git-svn-id: http://svn.topazproject.org/svn/head@716 48088309-b40f-0410-b149-b14953253675