Closes #174: Ensure uniqueness of authentication-ids .
Note that due to limitations with triplestores, simultaneous requests using the
same auth-id may get both rejected. Also, there are possible (very rare) error
cases where the database ends up with duplicate auth-ids anyway: this is because
due to the lack of locks and uniqueness constraints we must commit the new ids
and then check for duplicates afterwards, and if duplicates are found we must
roll back "manually" - this manual rollback could fail.


git-svn-id: http://svn.topazproject.org/svn/head@760 48088309-b40f-0410-b149-b14953253675