Assign implied permissions also during grant/revoke and cancel-grant/cancel-revoke operations.
Implied permissions are typically permissions required to access
services that topaz depends on and may have an external direct
exposure. eg. Fedora API-A REST calls will be made by PLoS to
retrieve annotation body, article content etc. The permissions
that Fedora exposes should mirror the permissions that topaz
uses; but without having to have Topaz client applications
set the permissions explicitly for them.

Note that the implied permissions apply to the same resource
that the client is grant/revoke-ing permissions on. If the 
resource that the dependent service uses for access verification
is different then the XACML policies should be made aware of 
how to do the resource mapping.

(Addresses #102)



git-svn-id: http://svn.topazproject.org/svn/head@875 48088309-b40f-0410-b149-b14953253675