Hopefully permanently remove need for [861] on fedora-install by windows users
Don't know why, but windows lucene wants its database initialized whereas
unix appears to do this for free. This work around here simply copies in
an initialized db before fedora is started. 

It may be better to do [861] but fedora needs to be started to do this and
it needs to happen only once after fedora is started the first time. This is
more difficult to work into our scripts (and might even require a modification
to fedora).

fedoragsearch should probably be modified to do a check on the existance of
the database in its context initializer. re #205.


git-svn-id: http://svn.topazproject.org/svn/head@879 48088309-b40f-0410-b149-b14953253675