Remove test broken by jdk 1.5_08.
It is unlikely something we will ever run into. IMO, it is a new bug
in the JDK itself. We could potentially work around it by spinning
a tread in LogManager's static initializer and trying to set 
Logger.global's parent class or adding a handler for it, but it
adds more complexity at this point for no value add.


git-svn-id: http://svn.topazproject.org/svn/head@931 48088309-b40f-0410-b149-b14953253675