Make it reasonably easy to startup mulgara as a separate server for testing
To do so, use the following:
  mvn -Pit-mulgara pre-integration-test

I spent an hour or two trying to get rid of the spurious warnings (that
display as errors) about logging mis-configuration. Their distracting,
but at this point I believe they're safe to ignore. Maybe somebody else
can take a shot at cleaning them up.

I also tried to avoid putting this in a profile. However, for some
reason maven is insisting on running the pre-integration-test phase 
when 'mvn install' is run. This of course would hang the build, so
I put stuff in a profile. That's maven.


git-svn-id: http://svn.topazproject.org/svn/head@959 48088309-b40f-0410-b149-b14953253675