Run mulgara in a separate container fro integration tests
This is related to [959]. It uses the same mechanism except it
uses the it-startenv profile and it does not block once mulgara's
container is started -- it keeps going so other things can test.
Also, it does its work in the "package" phase. This is because
this is the phase immediately before pre-integration-test where
fedora is started (and mulgara needs to be up before fedora is
started).

The default URI for mulgara has been changed to:
  http://localhost:9091/mulgara-service-0.5-SNAPSHOT/services/...


git-svn-id: http://svn.topazproject.org/svn/head@961 48088309-b40f-0410-b149-b14953253675