Fix minor maven build issue
Build was broken by [925] if done on a clean account (no .m2/repository).
It is hard to explain what exactly is the problem other than there is
probably a maven bug somewhere that discusses how ${pom.parent.version}
is broken. It is probably best to just leave the <version> tag out and
the parent version number is probably inherited that way.

mvn merge -r925:924 pom.xml


git-svn-id: http://svn.topazproject.org/svn/head@973 48088309-b40f-0410-b149-b14953253675