fixes #201 Expand feed to support true categories
The feed for artices AND alerts has been EXPANDED as such:
 * What used to be tagged as <categories><category>... is now
   tagged as <subjects><subject>... and contains the unparsed
   category/sub-category as before.
 * Parsed categories are now supported as <categories><category>...
   and represent the <topaz:mainCategory> as designated in [772].

In the future, sub-category support should be added for 
completeness. It is available today by subtracting the category
from the subject, but that's clearly suboptimal. Doing yet more
queries today for it, on the other hand, will slow down processing.
There is probably a way to do with with sub-queries that should
be invesitgated.


git-svn-id: http://svn.topazproject.org/svn/head@1190 48088309-b40f-0410-b149-b14953253675