Hacks to work around bug http://jira.codehaus.org/browse/MNG-1412 and the fact that mulgara.jar includes an old version of log4j.
Specifically, we need log4j 1.2.12 or later; however, the mulgara jar
contains log4j classes but from version 1.2.8 . So we need to ensure that
the log4j jar preceeds the mulgara jar on the classpath. But due to bug
MNG-1412 the order of the jars on the classpath is pretty random, instead
of as listed in the pom (including being different depending on whether
you run 'mvn install' or 'mvn clean install'). Adding these two exclusions
seems to ensure the log4j jar always preceeds the mulgara jar.


git-svn-id: http://svn.topazproject.org/svn/head@1349 48088309-b40f-0410-b149-b14953253675