Fix rollbackTxn() to always properly reset the session.
Mulgara requires a setAutoCommit(true) to reset the sesison after an error.
While ItqlInterpreterBean.rollback() calls 'rollback(); setAutoCommit(true)',
that is not sufficient because rollback() will throw an exception if the
transaction failed, so the setAutoCommit(true) doesn't get called in that
case. But commit() will always work (but will actually roll back the
transaction if it had failed) since that just calls setAutoCommit(true).

Note things are not bulletproof because we don't check if the commit() fails
due to a problem communicating with the server.


git-svn-id: http://svn.topazproject.org/svn/head@1380 48088309-b40f-0410-b149-b14953253675