Attempt to detect CAS PGT expiry and try to handle the situation gracefully.
This requires the CASProtectedService to hold on to the user's HTTPSession.
Therefore session-maps are no longer supported by CASProtectedService.

If a proxy ticket cannot be obtained based on a PGT, the
assumption is that it is because the PGT is invalid. So
the recovery strategy is to remove the CASReceipt from
the user's HTTPSession and additionally throwing a 
special exception (NoProxyTicketException) to the caller.

So apps that hold service tickets (like the Plos-app) can catch this exception
may attempt a revalidate with CAS that may or may not result in the user 
having to login again. (based on the user's TGC expiry status)

Apps that hold proxy tickets (like the Topaz-apps) can let the caller 
handle that. If the caller is using AbstractReAuthStubFactory 
based stub instances, then the automatic re-auth should take
care of this. (See the addition to AxisAuthFaultDetector)



git-svn-id: http://svn.topazproject.org/svn/head@1439 48088309-b40f-0410-b149-b14953253675