This should fix the CAS problems we were seeing.  This checkin does two things:
1) It invalidates the existing user session if the user logs in via CAS and sets
up a new session with all the old session attributes except the PLoS objects.
Note that jetty is now upgraded to 6.0.2 to allow this to function correctly 
because of a bug in prior versions: http://jira.codehaus.org/browse/JETTY-141

2) Properly use the session scoped beans in Spring by setting <aop:scoped-proxy/> for
each bean that is session scoped so that a proxy object created to access it by 
objects which are not session scoped but which use these session scoped beans.  
The EnsureUserAccountInterceptor is one such object (and which was causing the 
problems we were seeing).  I have also upgraded to using the latest Spring release.



git-svn-id: http://svn.topazproject.org/svn/head@1443 48088309-b40f-0410-b149-b14953253675