Created a tools directory to create a bunch of useful tools/applications to be used external to PLoS ONE application.
Added the first application which is a command line utility to extract various
alert associated information from Mulgara directly. This utility can be
executed at the 'head' level as:

    mvn -f topazproject/tools/pom.xml -DAlertJournals \
           -Dcommand.line="-uri <mulgara uri>"

The pom.xml of 'alerts' uses profiles to allow execution of multiple classes
within the jar. Please refer to the file and copy over the <profile> to a new
one and modify the <name> and <mainClass>.


git-svn-id: http://svn.topazproject.org/svn/head@1476 48088309-b40f-0410-b149-b14953253675