ant-tasks:svninfo task puts svn info into build so we can know what is deployed
As operations is deploying directly out of subversion, when problems happen
it is important to know what version is deployed. This task will create a file
that can be placed anywhere in a web-application, jar-file, etc.

For example, to put it in the root of a webapp (such that it would be fetchable
via an http request), add the following to the webapps pom.xml:
{{{
<plugin>
  <groupId>org.topazproject.plugins</groupId>
  <artifactId>ant-tasks</artifactId>
  <executions>
    <execution>
      <phase>generate-resources</phase>
      <goals><goal>svninfo</goal></goals>
      <configuration>
        <directory>..</directory>
        <outputPath>${project.build.directory}/${project.build.finalName}/svninfo.txt</outputPath>
      </configuration>
    </execution>
  </executions>
</plugin>
}}}

Be sure you have the <diretory> you want svninfo on. It might be above the
directory where your project is.


git-svn-id: http://svn.topazproject.org/svn/head@1492 48088309-b40f-0410-b149-b14953253675