Add some more configuration to alerts
Primary issue is making sure that we use port 2525 for integration tests and 
port 25 in production. For now the default is still 2525 (because I can't run
integration tests myself) and we're putting port 25 into the default topaz.xml
configuration file.

Down the road, the email.xsl also needs to move out of WEB-INF/classes. We
can do this the same way we specify the location of the mulgara and lucene
databases. 


git-svn-id: http://svn.topazproject.org/svn/head@2067 48088309-b40f-0410-b149-b14953253675