Improve blind search for jdk in /opt
This was an attempt automatically find java. Should probably be
moved into one place: /etc/sysconfig/topaz-common, but then one
specific rpm would have to create that (tools?).

This fix calls -maxdepth first (so find doesn't complain) and 
makes sure we find a directory (and not a tar file as we did
earlier today).


git-svn-id: http://svn.topazproject.org/svn/head@2166 48088309-b40f-0410-b149-b14953253675