If all the results fit on one page, return the correct total number of results
One issue: If the next page has no results, you may get an empty page.

The reason we do all this is because xacml may be filtering our results
and the number of results we return normally (without this optimization)
is the number from lucene (unfiltered by xacml). So the number we return
is usually just an approximation.


git-svn-id: http://svn.topazproject.org/svn/head@2185 48088309-b40f-0410-b149-b14953253675