Filter search results by user
This allows admin to see all articles (not just published articles).

This was really a bug as search was using xacml as user anonymous.
So no user filtering was going on and no customization of filtering
would be possible in the future without this.

The plos frontend was changed to use search in session scope instead
of application scope.

Thanks Viru & Pradeep.


git-svn-id: http://svn.topazproject.org/svn/head@2186 48088309-b40f-0410-b149-b14953253675