Enable src/{main,test,it}/groovy (if there's code there)
This just means that individual pom files do not need to be modified to
support groovy. Just put your code in the appropriate location and it will
build. (Though you may need to add groovy:groovy-all:1.0 as a dependency.)

Note that due to MSUREFILRE-121, integration tests are a problem because
dependencies get lost. The work around is to add
  <properties><forkMode>never</forkMode></properties>
This is fine for individual modules, but not for the integraitontests
module, so beware. I have not modified the integrationtests module to
pull in src/it/groovy. It is rather a shame as using groovy for testing
is semi interesting. Of course so is testNG that is now available that
we're using java 1.5. The integrationtests module is a terrible mvn hack
anyway IMO -- at least the way it is now.


git-svn-id: http://svn.topazproject.org/svn/head@2281 48088309-b40f-0410-b149-b14953253675