Remove stuff from [2281] that broke build due to maven bugs
I got pretty far along a maven-plugin to fix this the right way, but it was
clearly going to take a bit longer.

According to Ronald, this is the issue: (also see [2287])

The problem is that maven does not properly "reload" plugins when
moving from pom to pom. So when you added the antrun stuff to the head
pom, maven starts loading the antrun plugin early on, but it uses the
definitions (dependencies) at that point; and when it reaches the
mulgara-service pom it doesn't reload the antrun plugin, and hence
doesn't provide the maven-artifact-ant lib to the plugin.

While the above error was an easy fix, another stupid
maven bug prevents the webservices from building: because the antrun
plugin is listed both in the head and the child poms, it appears twice
in the effective-pom, and hence the tasks are all run twice. This is
mostly annoying in a number of cases, but in the case of the
webservice impl's where the CopyServiceDeploymentDescriptors copies
stuff out of the generated-sources directory and then deletes that
directory, the second run fails because, well, we've already deleted
that directory.

Agh! Stupid, stupid, stupid. And they won't fix it in the 2.0 branch:
http://www.mail-archive.com/issues@maven.apache.org/msg16142.html

Also see #95 (we've seen some of these issues before)


git-svn-id: http://svn.topazproject.org/svn/head@2291 48088309-b40f-0410-b149-b14953253675