New fedora package using our tomcat and fedora 2.2 war file
This depends on the following manually created artifacts on gandalf:
 * info.fedora:fedora:2.2:war from [2342]
 * info.fedora:fedorahome:2.2:zip from [2343]
 * info.fedora:fedora-server:2.2:jar from [2334]

This is not yet ready for prime time. I'm checking in the (more or less)
stock configuration files so that diffs in the future will make more sense.
The config files that will be changed in a subsequent checkin are:
 * fedora.fcfg (initial version generated via 2.2 installer quick install)
 * log4j.properties (comes from same method as fedora.fcfg)
 * server.xml (copied from packages/topaz -- the one generated by the fedora
   2.2 installer is inappropriate because it is for a different version of
   tomcat)


git-svn-id: http://svn.topazproject.org/svn/branches/fedora22branch@2344 48088309-b40f-0410-b149-b14953253675