Configuration (and minor packaging) changes for fedora 2.2 to work
Changed the following configuration files:
 * log4j.properties - Put logs into /var/log/topaz and rotate the way we like
 * server.xml - Listen on 9090, 9443 & 9005
 * fedora.fcfg
   * Store data in /var/lib/topaz/data/fedora
   * Use postgresql
   * Specify the ports we use in server.xml

To upgrade to postgresql, just make sure postgresql is installed. Then
follow the instructions at:
  http://www.fedora.info/download/2.2/userdocs/server/cmd-line/index.html#rebuild

(I tried to do this, but got mucked up in other stuff I was testing. So
this needs to be tested thoroughly before deployment.)

Packaging issues: (to be fixed when we patch assembly plugin)
 * The assembly problem bites again. Even the bug_workaround tricks we tried
   in other packages were no help to allow me to replace the log4j.properties
   in fedorahome.zip. So the RPM puts it in the right place.


git-svn-id: http://svn.topazproject.org/svn/branches/fedora22branch@2345 48088309-b40f-0410-b149-b14953253675